pyproject.toml
[tool.poetry]
name = "ai-vtuber"
version = "0.30.0"
description = "AI Vtuber，结合ASR、LLM、TTS、SVC、SD等功能为一体的开源AI软件"
authors = ["ikaros <327209194@qq.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "2.32.3"
wenxinworkshop = { git = "https://gitee.com/ikaros-521/WenxinWorkshop-Python-SDK", rev = "12b863fca7ec1cdf969d73a04b50f66d2d24a878" }
blivedm = { git = "https://gitee.com/ikaros-521/blivedm", rev = "c9ac671f783c4c3eaa4ee4e7738226ebbf402259" }
sparkdesk-api = { git = "https://gitee.com/ikaros-521/sparkdesk-api", rev = "bb830c0f1a1b5c575a29a24c4077d6482aa9e6e1" }
openai = "*"
pygame = "*"
edge-tts = "6.1.6"
langid = "*"
aiohttp = "3.9.1"
asyncio = "*"
slack_sdk = "3.21.3"
profanity = "1.1"
elevenlabs = "0.2.8"
PyPDF2 = "*"
protobuf_inspector = "*"
websocket-client = "*"
websocket-client-py3 = "*"
tiktoken = "*"
pyahocorasick = "*"
webuiapi = "*"
pyvirtualcam = "*"
numpy = "*"
faiss-cpu = "*"
sentence_transformers = "*"
pydub = "*"
pydantic = "1.10.13"
send2trash = "*"
keyboard = "*"
PyAudio = "*"
SpeechRecognition = "*"
baidu-aip = "*"
chardet = "*"
pypinyin = "*"
gradio_client = "==0.8.1"
schedule = "*"
bilibili-api-python = "*"
langchain = "0.0.142"
playwright = "*"
protobuf = "3.20.3"
certifi = "*"
charset-normalizer = "*"
greenlet = "*"
idna = "*"
lxml = "*"
pyee = "*"
typing_extensions = "*"
urllib3 = "2.2.2"
pyhandytools = "*"
curl_cffi = "*"
zhipuai = "*"
bardapi = "*"
revTongYi = "*"
pyautogui = "*"
nicegui = "1.4.30"
pytchat = "*"
PySocks = "*"
pure-protobuf = "3.0.0a5"
xingchen = "1.1.5"
emoji = "*"
google-generativeai = "*"
colorlog = "*"
faster_whisper = "*"
httpx = "0.25.2"
Pillow = "10.1.0"
pygtrans = "*"
jieba = "*"
TikTokLive = "*"
azure-cognitiveservices-speech = "*"
pyjwt = "*"
dashscope = "*"
pygetwindow = "*"
opencv-python = "*"
g4f = "*"
starlette = "0.36.3"
pyserial = "*"

# dy2
betterproto = "2.0.0b6"
loguru = "*"

# asr
funasr = "1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
